Cache Simulator Characteristics
The simulator will emulate accesses to data only (Data).
The simulator will include two levels of cache (L1 and L2).
Both caches will operate under the No Write Allocate or Write Allocate policy, with a Write Back policy.
The cache memory will operate based on the inclusion principle (Inclusive).
The eviction policy is LRU (Least Recently Used).
At the start of the simulation, the cache memory is empty.
All accesses are 4 bytes in size and aligned to a 4-byte boundary (the two least significant bits of the address will always be 00).